The updates to the README need more polish.
---

Found a fairly glaring error in the `gene_region_finder.sh` logic where the sorting was being done as `k6,6n` which is incorrect.

The 6th field has the LD and sorting by the LD gives you the wrong start and stop points.
The start point is the first BP with within the LD cutoff and the stop point is the last BP within the LD cutoff.
When you sort by LD, the first and the last lines will have the extremes of the LD (within the cutoff you chose) but the first and the last line will not necessarily have the extremes of the BPs.
For example, when you have an LD cutoff of 0.4 and get the following hypothetical data

LD BP
0.4 12
0.5 11
0.6 17
0.7 13

what you what is BP 11 as your start and BP 17 as your stop but as you can see sorting by LD will give 12 and 13 which is incorrect. This is because the Highest LD does not necessaily associate with the last BP and the lowest LD does not necessarily associate with the lowest LD.

This commit fixes this error
---